[runtime] Move continuation/awaitable maps from DEM to ATCM
Address xintongsong's review feedback (#546 comments #6, #7): the
continuationContexts and pythonAwaitableRefs maps are runner-context
concerns, not durable-execution concerns, and their placement in
DurableExecutionManager forced ActionTaskContextManager to hold a
manager-to-manager reference forbidden by 545-DESIGN.md.

Move both maps and their accessors into ActionTaskContextManager.
actionTaskDurableContexts stays in DEM since DEM consumes it via
setupDurableExecutionContext. Drop the DEM parameter from
createAndSetRunnerContext (no longer needed); transferContexts keeps
the DEM parameter for putDurableContext only.
